# backend/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Distributed Job Scheduler API
  description: High-performance, multi-tenant distributed job processing platform.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: /api/v1
    description: Base API URL
paths:
...
  /auth/refresh:
    post:
      operationId: refreshToken
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New tokens generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
        '401':
          description: Invalid or expired refresh token
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RefreshRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          example: eyJ...
        deviceName:
          type: string
          example: Chrome Windows
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: eyJ...
        refreshToken:
          type: string
          example: eyJ...
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        isEmailVerified:
          type: boolean
...